@@ -21,63 +21,63 @@
 		"node": ">=16.0.0"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.14.0",
-		"@babel/plugin-proposal-class-properties": "^7.13.0",
-		"@babel/plugin-proposal-object-rest-spread": "^7.13.8",
-		"@babel/plugin-transform-runtime": "^7.13.15",
-		"@babel/preset-env": "^7.14.0",
-		"@babel/preset-react": "^7.13.13",
-		"@babel/preset-typescript": "^7.13.0",
-		"@storybook/addon-actions": "^6.2.9",
-		"@storybook/addon-essentials": "^6.2.9",
-		"@storybook/addon-links": "^6.2.9",
+		"@babel/core": "^7.18.13",
+		"@babel/plugin-proposal-class-properties": "^7.18.6",
+		"@babel/plugin-proposal-object-rest-spread": "^7.18.9",
+		"@babel/plugin-transform-runtime": "^7.18.10",
+		"@babel/preset-env": "^7.18.10",
+		"@babel/preset-react": "^7.18.6",
+		"@babel/preset-typescript": "^7.18.6",
+		"@storybook/addon-actions": "^6.5.10",
+		"@storybook/addon-essentials": "^6.5.10",
+		"@storybook/addon-links": "^6.5.10",
 		"@storybook/addon-postcss": "^2.0.0",
-		"@storybook/react": "^6.2.9",
+		"@storybook/react": "^6.5.10",
 		"@types/classnames": "^2.3.1",
-		"@types/d3": "^6.3.0",
+		"@types/d3": "^7.4.0",
 		"@types/query-string": "^6.3.0",
-		"@types/react": "^17.0.4",
-		"@types/react-dom": "^17.0.3",
-		"@types/react-redux": "^7.1.16",
-		"@types/react-router-dom": "^5.1.7",
-		"@typescript-eslint/eslint-plugin": "^4.22.0",
-		"@typescript-eslint/eslint-plugin-tslint": "^4.22.0",
-		"@typescript-eslint/parser": "^4.22.0",
-		"assets-webpack-plugin": "^7.1.0",
-		"autoprefixer": "^10.2.5",
-		"babel-loader": "^8.2.2",
-		"babel-plugin-macros": "^3.0.1",
-		"clean-webpack-plugin": "^3.0.0",
-		"core-js": "^3.11.1",
-		"css-loader": "^5.2.4",
-		"css-minimizer-webpack-plugin": "^2.0.0",
-		"eslint": "^7.25.0",
+		"@types/react": "^18.0.17",
+		"@types/react-dom": "^18.0.6",
+		"@types/react-redux": "^7.1.24",
+		"@types/react-router-dom": "^5.3.3",
+		"@typescript-eslint/eslint-plugin": "^5.35.1",
+		"@typescript-eslint/eslint-plugin-tslint": "^5.35.1",
+		"@typescript-eslint/parser": "^5.35.1",
+		"assets-webpack-plugin": "^7.1.1",
+		"autoprefixer": "^10.4.8",
+		"babel-loader": "^8.2.5",
+		"babel-plugin-macros": "^3.1.0",
+		"clean-webpack-plugin": "^4.0.0",
+		"core-js": "^3.25.0",
+		"css-loader": "^6.7.1",
+		"css-minimizer-webpack-plugin": "^4.0.0",
+		"eslint": "^8.23.0",
 		"eslint-config-prettier": "^8.5.0",
-		"eslint-webpack-plugin": "^2.5.4",
-		"fork-ts-checker-webpack-plugin": "^6.2.5",
-		"mini-css-extract-plugin": "^1.5.1",
-		"postcss": "^8.2.13",
-		"postcss-loader": "^5.2.0",
+		"eslint-webpack-plugin": "^3.2.0",
+		"fork-ts-checker-webpack-plugin": "^7.2.13",
+		"mini-css-extract-plugin": "^2.6.1",
+		"postcss": "^8.4.16",
+		"postcss-loader": "^7.0.1",
 		"prettier": "^2.7.1",
-		"sb": "^6.2.9",
-		"terser-webpack-plugin": "^5.1.1",
-		"tslib": "^2.2.0",
+		"sb": "^6.5.10",
+		"terser-webpack-plugin": "^5.3.5",
+		"tslib": "^2.4.0",
 		"tslint": "^6.1.3",
-		"typescript": "^4.2.4",
-		"webpack": "^5.36.1",
-		"webpack-bundle-analyzer": "^4.4.1",
-		"webpack-cli": "^4.6.0"
+		"typescript": "^4.8.2",
+		"webpack": "^5.74.0",
+		"webpack-bundle-analyzer": "^4.6.1",
+		"webpack-cli": "^4.10.0"
 	},
 	"dependencies": {
-		"@babel/runtime": "^7.14.0",
-		"@reduxjs/toolkit": "^1.5.1",
+		"@babel/runtime": "^7.18.9",
+		"@reduxjs/toolkit": "^1.8.5",
 		"classnames": "^2.3.1",
-		"d3": "^6.7.0",
-		"query-string": "^7.0.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-redux": "^7.2.4",
-		"react-router-dom": "^5.2.0",
+		"d3": "^7.6.1",
+		"query-string": "^7.1.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-redux": "^8.0.2",
+		"react-router-dom": "^6.3.0",
 		"whatwg-fetch": "^3.6.2"
 	},
 	"packageManager": "yarn@3.2.3"
