@@ -1,25 +1,24 @@
 import React, { useEffect } from 'react';
-import { useDispatch, useSelector, batch } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { Home } from '~/pages';
-import { IWeightModel } from '~/models';
-import { IState, weightsDuck, profileDuck } from '~/redux';
+import { IState, weightsSlice, profileSlice } from '~/redux';
 
 export default function HomeContainer() {
 	const dispatch = useDispatch();
+	const { loadAllWeights, saveWeight } = bindActionCreators(
+		weightsSlice.actions,
+		dispatch
+	);
+	const { getProfile } = bindActionCreators(profileSlice.actions, dispatch);
 	const weights = useSelector((state: IState) => state.weights);
 	const profile = useSelector((state: IState) => state.profile.profile);
 	const today = useSelector((state: IState) => state.diagnostics.today);
 
 	useEffect(() => {
-		batch(() => {
-			dispatch(weightsDuck.actions.loadAllWeights());
-			dispatch(profileDuck.actions.getProfile());
-		});
+		loadAllWeights();
+		getProfile();
 	}, []);
-
-	const handleWeightSaved = (weight: IWeightModel) => {
-		dispatch(weightsDuck.actions.saveWeight(weight));
-	};
 
 	return (
 		<Home
@@ -28,7 +27,7 @@ export default function HomeContainer() {
 				profile,
 				today
 			}}
-			onSaveWeight={handleWeightSaved}
+			onSaveWeight={saveWeight}
 		/>
 	);
 }
