@@ -1,7 +1,7 @@
 export { default as IState } from './IState';
 export { createStore } from './store';
-export * as errorDuck from './error.duck';
-export * as authDuck from './auth.duck';
-export * as diagnosticsDuck from './diagnostics.duck';
-export * as weightsDuck from './weights.duck';
-export * as profileDuck from './profile.duck';
+export { default as errorSlice } from './error.slice';
+export { default as authSlice } from './auth.slice';
+export { default as diagnosticsSlice } from './diagnostics.slice';
+export { default as weightsSlice } from './weights.slice';
+export { default as profileSlice } from './profile.slice';
